Remove size_t in preference of uint32_t

+++ c/inc/codec/codec.h
@@ -15,18 +15,18 @@ HexChar_Decode(const char hexHigh, const char hexLow, uint8_t* byte);
 CryptoResult
 Base64_Encode(
     const uint8_t* bytes,
-    const size_t byteCount,
+    const uint32_t byteCount,
     uint8_t* b64Buf,
-    const size_t bufSize,
-    size_t* required);
+    const uint32_t bufSize,
+    uint32_t* required);
 
 CryptoResult
 Base64_Decode(
     const uint8_t* b64Bytes,
-    const size_t b64Count,
+    const uint32_t b64Count,
     uint8_t* byteBuf,
-    const size_t bufSize,
-    size_t* required);
+    const uint32_t bufSize,
+    uint32_t* required);
 
 #endif
